+++ gaios/Shared/Controllers/DialogLoginPassphraseViewController.swift
@@ -113,7 +113,7 @@ class DialogLoginPassphraseViewController: KeyboardViewController {
     }
 
     @IBAction func btnLearnMore(_ sender: Any) {
-        UIApplication.shared.open(ExternalUrls.passphraseReadMore, options: [:], completionHandler: nil)
+        SafeNavigationManager.shared.navigate(ExternalUrls.passphraseReadMore)
     }
 
     @IBAction func passphraseDidChange(_ sender: Any) {